Update prefabs, assign and place monsters at beginning of game

+++ DejarikLibrary/BoardGraph.cs
@@ -89,8 +89,8 @@ namespace DejarikLibrary
             double x = Math.Cos(angle) * ((innerBoundingRadius + outerBoundingRadius) / 2.0);
             double y = Math.Sin(angle) * ((innerBoundingRadius + outerBoundingRadius) / 2.0);
 
-            node.XPosition = x;
-            node.YPosition = y;
+            node.XPosition = (float)x;
+            node.YPosition = (float)y;
         }
 
         private Dictionary<NodeMapKey, List<NodePath>> BuildNodeMap(List<Node> nodes)
